badgeButton refatored
